Save the current patchset in the kernel
In the good old days when only a small number of computer boards were
released per year and mainline kernel support was available for most of
them, it was an excellent idea to save the current kernel configuration
in the binary kernel. This made it possible to reproduce a kernel from
the vanilla kernel sources at any time.

In modern times, however, when myriads of (mostly ARM based) computer
boards are inundating the market every year most of which require a
large patchset that often is difficult to obtain, the saved config file
is nearly useless. (If a patchset cannot be obtained with reasonable
effort, the board vendor, of course, is in breach of the Linux kernel
license.)

This patch now provides a mechanism to store the current patchset in the
binary kernel in a very similar way as it is done with the config file. To
reproduce the source tree that was used for a particular kernel, use the
command sequence:
  tar zxf /proc/patchset.tar.gz baseversion
  major=`cut -d. -f1 baseversion`
  urldir=http://www.kernel.org/pub/linux/kernel/v$major.x
  dir=linux-`cat baseversion`
  rm -f baseversion
  archive=$dir.tar.xz
  wget $urldir/$archive
  tar Jxf $archive
  cd $dir
  tar zxf /proc/patchset.tar.gz
  quilt push -a
  zcat /proc/config.gz >.config

BTW: Making the kernel sources available through this mechanism does not
entirely fulfill the disclosure obligations of the GPL, but it certainly
is much better than making available nothing.

Signed-off-by: Carsten Emde <C.Emde@osadl.org>