fs: aufs: fix a build error for archs which doesn't support MUTEX_SPIN_ON_OWNER
For struct mutex, the member "owner" only exist when either
DEBUG_MUTEXES or MUTEX_SPIN_ON_OWNER is enabled. The SMP is a necessary
condition, but not a sufficiency condition for MUTEX_SPIN_ON_OWNER.
So we would get build error on arch such as mpis due to SMP is
enabled and both DEBUG_MUTEXES and MUTEX_SPIN_ON_OWNER are disabled.
Replace the SMP with MUTEX_SPIN_ON_OWNER to fix it.

Signed-off-by: Kevin Hao <kexin.hao at windriver.com>
Signed-off-by: Pavel Pisa <ppisa@pikron.com>